Update fixtures to support Fuel-10.0
- Updated scripts to generate fixtures for 10.0
- Script now generates fake nodes to work with
- Added multirack support and fixtures
- Uploaded new fixtures 10.0 build #260

Partial-bug: #1535339

+++ utils/generate_yamls.sh
@@ -1,12 +1,45 @@
 #!/bin/bash
 #
 # Use this script to generate and save astute.yaml fixtures.
-# Should be executed on Fuel node with at least 7 discovered
-# and unused (not assigned to any env) nodes.
-#
+# Should be executed on Fuel node with 'advanced' feature enabled
+# (see FEATURE_GROUPS list in /etc/nailgun/settings.yaml)
+
+CWD=$(cd `dirname $0` && pwd -P)
 
 mkdir ./yamls
 rm -f ./yamls/*
+
+function generate_fake_nodes_fixtures {
+  # $1 - first IP of admin network to start generate nodes from
+  # $2 - number of nodes to generate
+  # $3 - name of fixture
+  $CWD/generate_nodes_fixtures.rb $1 $2 > $CWD/fixtures/${3}.json
+}
+
+function create_fake_nodes {
+  manage.py loaddata $CWD/fixtures/${1}.json
+}
+
+function clean_fake_nodes {
+  fuel nodes | grep -q 'discover | fnode-' &&
+    fuel nodes | awk '/discover \| fnode-/{ print $1 }' | xargs fuel node --delete-from-db --force --node
+}
+
+function admin_net_tpl {
+  fuel network-group list | awk '/^1 /{print $9}' | sed -e 's/\.[[:digit:]]\+$//'
+}
+
+function id_of_role {
+  env=$1
+  role=$2
+  yaml=`grep -rl node_roles: deployment_$env/*yaml | head -n1`
+  ruby -ryaml -e '
+  astute = YAML.load(File.read(ARGV[0]))
+  role = ARGV[1]
+  node = astute["network_metadata"]["nodes"].find{|key, hash| hash["node_roles"].include?("#{role}") }
+  puts node.last["uid"]
+  ' $yaml $role
+}
 
 function enable_ceph {
   fuel env --attributes --env $1 --download
@@ -18,6 +51,16 @@ function enable_ceph {
   attr["editable"]["storage"]["ephemeral_ceph"]["value"] = true
   attr["editable"]["storage"]["volumes_lvm"]["value"] = false
   attr["editable"]["storage"]["osd_pool_size"]["value"] = "2"
+  File.open(ARGV[0], "w").write(attr.to_yaml)' "cluster_$1/attributes.yaml"
+  fuel env --attributes --env $1 --upload
+  rm -rf "cluster_$1"
+}
+
+function enable_cblock {
+  fuel env --attributes --env $1 --download
+  ruby -ryaml -e '
+  attr = YAML.load(File.read(ARGV[0]))
+  attr["editable"]["storage"]["volumes_block_device"]["value"] = true
   File.open(ARGV[0], "w").write(attr.to_yaml)' "cluster_$1/attributes.yaml"
   fuel env --attributes --env $1 --upload
   rm -rf "cluster_$1"
@@ -88,7 +131,12 @@ function enable_vms_conf {
 }
 
 function list_free_nodes {
-  fuel nodes 2>/dev/null | grep discover | grep None | awk '{print $1}'
+  # list unused nodes from the list of fake nodes
+  if [ -n "$1" ] ; then
+    fuel nodes 2>/dev/null | grep discover | grep None | grep 'fnode-' | grep $1 | awk '{print $1}'
+  else
+    fuel nodes 2>/dev/null | grep discover | grep None | grep 'fnode-' | awk '{print $1}'
+  fi
 }
 
 function save_yamls {
@@ -100,10 +148,23 @@ function envid {
   fuel env 2>/dev/null | grep $1 | awk '{print $1}'
 }
 
+function fix_node_names {
+  file=$1
+  ruby -ryaml -e '
+  astute = YAML.load(File.read(ARGV[0]))
+  astute["network_metadata"]["nodes"].each do |key, hash|
+    wrong = hash["name"]
+    puts "\"s/#{wrong}/#{key}/g\""
+  end
+  ' $file | xargs -I {} sed -e {} -i $file
+}
+
 function store_yamls {
   for role in $3 ; do
-    src=`ls deployment_$1/${role}_*.yaml | head -n1`
+    id=`id_of_role $1 $role`
+    src="deployment_$1/${id}.yaml"
     cp $src ./yamls/$2-$role.yaml
+    fix_node_names ./yamls/$2-$role.yaml
   done
 }
 
@@ -111,6 +172,10 @@ function generate_yamls {
   env=`envid $1`
   name=$2
   roles=($3)
+
+  # Create fake nodes for our fixtures
+  generate_fake_nodes_fixtures $admin_first_ip 10 default_nodegroup
+  create_fake_nodes default_nodegroup
 
   if [ "${name/ceph}" != "$name" ] ; then
     enable_ceph $env
@@ -133,9 +198,22 @@ function generate_yamls {
   if [ "${name/public_ssl}" != "$name" ] ; then
     enable_public_ssl $env
   fi
+  if [ "${name/cblock}" != "$name" ] ; then
+    enable_cblock $env
+  fi
+
+  if [ "${name/multirack}" != "$name" ] ; then
+    # move controllers to custom node group
+    for id in `list_free_nodes 9.9.9` ; do
+      if [ "${roles[0]}" = "controller" ] ; then
+        fuel --env $env node set --node $id --role ${roles[0]}
+        roles=("${roles[@]:1}")
+      fi
+    done
+  fi
 
   for id in `list_free_nodes` ; do
-    if ! [ -z "${roles[0]}" ] ; then
+    if [ -n "${roles[0]}" ] ; then
       fuel --env $env node set --node $id --role ${roles[0]}
       roles=("${roles[@]:1}")
       sleep 1
@@ -153,11 +231,28 @@ function generate_yamls {
 
 function clean_env {
   env=`envid $1`
-  fuel env --delete --env $env
-  rm -rf "cluster_$1"
-  rm -rf "deployment_$env"
-  sleep 80
+  if fuel env --env $env | grep $1 ; then
+    fuel env --delete --env $env
+    rm -rf "cluster_$env"
+    rm -rf "deployment_$env"
+    sleep 80
+  fi
+  clean_fake_nodes
 }
+
+function add_nodegroup {
+  env=`envid $1`
+  name=$2
+
+  fuel --env $env nodegroup --create --name $name
+}
+
+clean_fake_nodes
+sleep 1
+
+# Get some context
+admin_net=$(admin_net_tpl)
+admin_first_ip="${admin_net}.100"
 
 # Neutron vlan ceph
 fuel env --create --name test_neutron_vlan --rel 2 --net vlan
@@ -166,7 +261,7 @@ clean_env 'test_neutron_vlan'
 
 # Neutron vlan addons
 fuel env --create --name test_neutron_vlan --rel 2 --net vlan
-generate_yamls 'test_neutron_vlan' 'neut_vlan.murano.sahara.ceil' 'controller controller compute mongo mongo cinder cinder-block-device' 'primary-controller controller compute primary-mongo mongo cinder cinder-block-device'
+generate_yamls 'test_neutron_vlan' 'neut_vlan.cblock.murano.sahara.ceil' 'controller controller compute mongo mongo cinder cinder-block-device' 'primary-controller controller compute primary-mongo mongo cinder cinder-block-device'
 clean_env 'test_neutron_vlan'
 
 # Neutron-dvr vlan
@@ -179,10 +274,10 @@ fuel env --create --name test_neutron_tun --rel 2 --net tun
 generate_yamls 'test_neutron_tun' 'neut_tun.ceph.murano.sahara.ceil' 'controller controller compute ceph-osd ceph-osd mongo mongo' 'primary-controller controller compute ceph-osd primary-mongo mongo'
 clean_env 'test_neutron_tun'
 
-# Neutron tun ironic
-fuel env --create --name test_neutron_tun --rel 2 --net tun
-generate_yamls 'test_neutron_tun' 'neut_tun.ironic' 'controller ironic' 'primary-controller ironic'
-clean_env 'test_neutron_tun'
+# Neutron vlan ironic
+fuel env --create --name test_neutron_vlan --rel 2 --net vlan
+generate_yamls 'test_neutron_vlan' 'neut_tun.ironic' 'controller ironic' 'primary-controller ironic'
+clean_env 'test_neutron_vlan'
 
 # Neutron-l3ha tun + nova_quota
 fuel env --create --name test_neutron_tun --rel 2 --net tun
@@ -194,7 +289,11 @@ fuel env --create --name test_neutron_tun --rel 2 --net tun
 generate_yamls 'test_neutron_tun' 'neut_tun.vms_conf' 'virt compute' 'virt'
 clean_env 'test_neutron_tun'
 
-# Neutron tun, addons, ceph, public and hotizon ssl
+# Multirack, Neutron tun, addons, ceph, public and horizon ssl
 fuel env --create --name test_neutron_tun --rel 2 --net tun
-generate_yamls 'test_neutron_tun' 'neut_tun.murano.sahara.ceil.public_ssl' 'controller controller mongo mongo compute ceph-osd ceph-osd' 'primary-controller compute ceph-osd primary-mongo'
+add_nodegroup 'test_neutron_tun' 'custom_group1'
+generate_fake_nodes_fixtures 9.9.9.150 5 custom_nodegroup
+create_fake_nodes custom_nodegroup
+generate_yamls 'test_neutron_tun' 'neut_tun.multirack.murano.sahara.ceil.ceph.public_ssl' 'controller controller controller mongo mongo compute ceph-osd ceph-osd' 'primary-controller compute ceph-osd primary-mongo'
 clean_env 'test_neutron_tun'
+
